refactor: remove settings button component
