+++ src/components/app-navigation/appNavigationStyles.ts
@@ -2,3 +2,11 @@ export const appNavigationStyles = {
   display: 'flex',
   justifyContent: 'center',
 };
+
+export const settingButtonStyles = {
+  backgroundColor: '#fff',
+  color: 'primary.main',
+  '&:hover': {
+    backgroundColor: '#fff',
+  },
+};
