--- src/components/buttons/settings-button/settingsButtonStyles.ts
@@ -1,7 +1,0 @@
-export const settingButtonStyles = {
-  backgroundColor: '#fff',
-  color: 'primary.main',
-  '&:hover': {
-    backgroundColor: '#fff',
-  },
-};
